+++ tests/Delay/VariableDelayLineFilter.cpp
@@ -19,30 +19,30 @@
 
 BOOST_AUTO_TEST_CASE( VariableDelayLineFilter_sinus_line16000_delaysinus_test )
 {
-  std::array<float, PROCESSSIZE> data;
+  std::array<double, PROCESSSIZE> data;
   for(ptrdiff_t i = 0; i < PROCESSSIZE; ++i)
   {
-    data[i] = std::sin(2 * boost::math::constants::pi<float>() * (i+1.)/48000 * 1000);
+    data[i] = std::sin(2 * boost::math::constants::pi<double>() * (i+1.)/48000 * 1000);
   }
-  std::array<float, PROCESSSIZE> datadelay;
+  std::array<double, PROCESSSIZE> datadelay;
   for(ptrdiff_t i = 0; i < PROCESSSIZE; ++i)
   {
     datadelay[i] = i * .25f;
   }
 
-  ATK::InPointerFilter<float> generator(data.data(), 1, PROCESSSIZE, false);
+  ATK::InPointerFilter<double> generator(data.data(), 1, PROCESSSIZE, false);
   generator.set_output_sampling_rate(48000);
-  ATK::InPointerFilter<float> generatordelay(datadelay.data(), 1, PROCESSSIZE, false);
+  ATK::InPointerFilter<double> generatordelay(datadelay.data(), 1, PROCESSSIZE, false);
   generatordelay.set_output_sampling_rate(48000);
 
-  std::array<float, PROCESSSIZE> outdata;
+  std::array<double, PROCESSSIZE> outdata;
 
-  ATK::VariableDelayLineFilter<float> filter(20000);
+  ATK::VariableDelayLineFilter<double> filter(20000);
   filter.set_input_sampling_rate(48000);
   filter.set_input_port(0, &generator, 0);
   filter.set_input_port(1, &generatordelay, 0);
 
-  ATK::OutPointerFilter<float> output(outdata.data(), 1, PROCESSSIZE, false);
+  ATK::OutPointerFilter<double> output(outdata.data(), 1, PROCESSSIZE, false);
   output.set_input_sampling_rate(48000);
   output.set_input_port(0, &filter, 0);
